Restructuracion de funciones, bt no implementado
